MUS-92 NoteBodyDrawerTest
fixed a testcase

+++ musicdroid/src/androidTest/java/org/catrobat/musicdroid/pocketmusic/test/note/draw/AbstractDrawerTest.java
@@ -86,7 +86,7 @@ public abstract class AbstractDrawerTest extends AndroidTestCase {
         assertEquals(expectedLine, canvas.getDrawnElements().poll());
     }
 
-    protected void assertCanvasElementQueueOval(int startX, int startY, int stopX, int stopY) {
+    protected void assertCanvasElementQueueOval(float startX, float startY, float stopX, float stopY) {
         String expectedLine = CanvasMock.createString(CanvasMock.DRAW_OVAL, startX, startY, stopX, stopY);
 
         assertEquals(expectedLine, canvas.getDrawnElements().poll());
